pyproject.toml
[tool.poetry]
name = "trivia"
version = "0.1.0"
description = "Question/Answer Game"
authors = ["pzerzis <pzerzis@wings-ict-solutions.eu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"
pytest = "^7.4.3"
faker = "^19.13.0"
pytest-factoryboy = "^2.6.0"
pytest-django = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
